Small update - Added CMakeUserPresets.json to .gitignore - Added variables you can pass to CMake or put in CMakeUserPresets.json to specify where your external checkers and ccache are.wq

+++ tictactoe_cli/tictactoe_cli.cpp
@@ -109,7 +109,7 @@ void runGame(int argc, const char** argv) {
 
 	while (true) {
 		std::string ix, iy;
-		int cx{}, cy{};
+		long cx{}, cy{};
 
 		const auto& current = board.getPlayer(playerTurn);
 		board.draw();
